Adds not found handling for delete operations
Improves delete operations in the application by adding a check for resource existence before attempting deletion.

This change ensures that a `NotFoundException` is thrown when attempting to delete a non-existent resource (Company, Role, or UserRole), providing better error handling and preventing unexpected behavior.

+++ API/src/modules/companies/companies.service.ts
@@ -1,9 +1,14 @@
-import { Injectable, ConflictException } from '@nestjs/common';
+import {
+	Injectable,
+	ConflictException,
+	NotFoundException,
+} from '@nestjs/common';
 import { InjectModel } from '@nestjs/sequelize';
 import { Companies } from './companies.model';
 import { v4 as uuidv4 } from 'uuid';
 import { CreateCompanyDto } from './dtos/create-company.dto';
 import { UpdateCompanyDto } from './dtos/update-company.dto';
+import DestroyedResponse from 'src/types/delete.response';
 
 @Injectable()
 export class CompaniesService {
@@ -54,11 +59,23 @@ export class CompaniesService {
 		});
 	}
 
-	delete(_id: string) {
-		return this.companiesModule.destroy({
+	async delete(_id: string): Promise<DestroyedResponse> {
+		const companyExists = await this.companiesModule.findByPk(_id);
+
+		if (!companyExists) {
+			throw new NotFoundException(`No company was found with id ${_id}`);
+		}
+
+		const hasDestroyed = await this.companiesModule.destroy({
 			where: {
 				_id,
 			},
 		});
+
+		return {
+			_id,
+			message: `Company with id ${_id} successfully deleted`,
+			success: hasDestroyed,
+		};
 	}
 }
