+++ API/src/modules/user-roles/user-roles.service.ts
@@ -1,4 +1,8 @@
-import { Injectable, ConflictException } from '@nestjs/common';
+import {
+	Injectable,
+	ConflictException,
+	NotFoundException,
+} from '@nestjs/common';
 import { InjectModel } from '@nestjs/sequelize';
 import { UserRoles } from './user-roles.model';
 import { v4 as uuidv4 } from 'uuid';
@@ -6,6 +10,7 @@ import { CreateUserRoleDto } from './dtos/create-user-role.dto';
 import { Op } from 'sequelize';
 import { UsersService } from '../users/users.service';
 import { RolesService } from '../roles/roles.service';
+import DestroyedResponse from 'src/types/delete.response';
 
 @Injectable()
 export class UserRolesService {
@@ -44,11 +49,23 @@ export class UserRolesService {
 		});
 	}
 
-	delete(_id: string): Promise<number> {
-		return this.userRolesModel.destroy({
+	async delete(_id: string): Promise<DestroyedResponse> {
+		const userRoleExists = await this.userRolesModel.findByPk(_id);
+
+		if (!userRoleExists) {
+			throw new NotFoundException(`No userRole was found with id ${_id}`);
+		}
+
+		const hasDestroyed = await this.userRolesModel.destroy({
 			where: {
 				_id,
 			},
 		});
+
+		return {
+			_id,
+			message: `userRole with id ${_id} successfully deleted`,
+			success: hasDestroyed,
+		};
 	}
 }
